+++ CHANGES.rst
@@ -3,6 +3,11 @@
 - Add support for column info on redshift late binding views
 - Add support for ``MAXFILESIZE`` argument to ``UNLOAD``.
   (`Issue #123 <https://github.com/sqlalchemy-redshift/sqlalchemy-redshift/issues/123>`_)
+- Add support for the `CREATE LIBRARY`_ command.
+  (`Issue #124 <https://github.com/sqlalchemy-redshift/sqlalchemy-redshift/issues/124>`_)
+
+.. _CREATE LIBRARY: https://docs.aws.amazon.com/redshift/latest/dg/r_CREATE_LIBRARY.html
+
 
 0.7.3 (2019-01-16)
 ------------------